pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "efqs"
version = "0.1.0"
description = "Extreme-Field QED Simulator for vacuum birefringence, pair production, and linearized gravitational coupling."
readme = "README.md"
requires-python = ">=3.8"
license = { file = "LICENSE" }
authors = [
    { name = "Arcticoder", email = "arcticoder@example.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.11",
    "matplotlib>=3.8",
    "pytest>=7.4",
    "PyYAML>=6.0",
]

[project.urls]
Homepage = "https://github.com/arcticoder/extreme-field-qed-simulator"
Issues = "https://github.com/arcticoder/extreme-field-qed-simulator/issues"

[tool.setuptools.packages.find]
where = ["src"]
